requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
dataclasses-json==0.6.7
datasets==4.8.4
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.1
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
gitpython==3.1.46
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.7.2
idna==3.11
iniconfig==2.3.0
instructor==1.14.5
jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.13
langchain-anthropic==1.4.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.21
langchain-ollama==1.0.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.19
mypy-extensions==1.1.0
narwhals==2.18.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
ollama==0.6.1
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pydeck==0.9.1
pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyyaml==6.0.3
ragas==0.4.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
scikit-network==0.33.5
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sqlalchemy==2.0.48
streamlit==1.55.0
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tomli==2.4.0
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uuid-utils==0.14.1
watchdog==6.0.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
